module editing validators
